Refactor creating SidebarListItems & indexEntries
