+++ packages/databyss-ui/components/SourcesContent/SourcesContent.js
@@ -1,6 +1,9 @@
 import React from 'react'
 import { Router } from '@reach/router'
-import { sortEntriesAtoZ } from '@databyss-org/services/sources/util'
+import {
+  sortEntriesAtoZ,
+  createIndexPageEntries,
+} from '@databyss-org/services/entries/util'
 import { SourceCitationsLoader } from '@databyss-org/ui/components/Loaders'
 import AuthorsContent from '@databyss-org/ui/components/SourcesContent/AuthorsContent'
 import AuthorCitations from '@databyss-org/ui/components/SourcesContent/AuthorCitations'
@@ -18,11 +21,13 @@ export const SourcesRouter = () => (
 const SourcesContent = () => (
   <SourceCitationsLoader>
     {sourceCitations => {
-      const sourcesData = Object.values(sourceCitations).map(value => ({
-        text: value.text,
-        citations: value.citations?.map(citation => citation.text?.textValue),
-        type: 'sources',
-      }))
+      const sourcesData = Object.values(sourceCitations).map(value =>
+        createIndexPageEntries({
+          text: value.text,
+          citations: value.citations?.map(citation => citation.text?.textValue),
+          type: 'sources',
+        })
+      )
 
       const sortedSources = sortEntriesAtoZ(sourcesData, 'text')
