+++ packages/databyss-ui/modules/Sidebar/routes/Pages.js
@@ -3,20 +3,23 @@ import { PagesLoader } from '@databyss-org/ui/components/Loaders'
 import {
   sortEntriesAtoZ,
   filterEntries,
-} from '@databyss-org/services/sources/util'
-import SidebarList from '../../../components/Sidebar/SidebarList'
+  createSidebarListItems,
+} from '@databyss-org/services/entries/util'
+import SidebarList from '@databyss-org/ui/components/Sidebar/SidebarList'
 
 const Pages = ({ filterQuery }) => (
   <>
     <PagesLoader>
       {pages => {
-        const _menuItems = Object.values(pages).map(p => ({
-          text: p.name,
-          type: 'pages',
-          route: '/pages',
-          id: p._id,
-          params: p._id,
-        }))
+        const _menuItems = Object.values(pages).map(p =>
+          createSidebarListItems({
+            text: p.name,
+            type: 'pages',
+            route: '/pages',
+            id: p._id,
+            params: p._id,
+          })
+        )
         // alphabetize list
         const sortedPages = sortEntriesAtoZ(_menuItems, 'text')
         const filteredEntries = filterEntries(sortedPages, filterQuery)
