+++ packages/databyss-ui/modules/Sidebar/routes/Topics.js
@@ -3,8 +3,9 @@ import { AllTopicsLoader } from '@databyss-org/ui/components/Loaders'
 import {
   sortEntriesAtoZ,
   filterEntries,
-} from '@databyss-org/services/sources/util'
-import SidebarList from '../../../components/Sidebar/SidebarList'
+  createSidebarListItems,
+} from '@databyss-org/services/entries/util'
+import SidebarList from '@databyss-org/ui/components/Sidebar/SidebarList'
 
 const topicsOverview = [
   {
@@ -17,13 +18,15 @@ const topicsOverview = [
 const Topics = ({ filterQuery }) => (
   <AllTopicsLoader>
     {topics => {
-      const topicsData = Object.values(topics).map(value => ({
-        text: value.text.textValue,
-        type: 'topics',
-        route: '/topics',
-        id: value._id,
-        params: value._id,
-      }))
+      const topicsData = Object.values(topics).map(value =>
+        createSidebarListItems({
+          text: value.text.textValue,
+          type: 'topics',
+          route: '/topics',
+          id: value._id,
+          params: value._id,
+        })
+      )
       const sortedTopics = sortEntriesAtoZ(topicsData, 'text')
       const filteredEntries = filterEntries(sortedTopics, filterQuery)
